Backend/requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
pymongo>=4.3.0
python-dotenv>=1.0.0
google-auth>=2.23.0
PyJWT>=2.8.0
plaid-python>=20.0.0